add: differentiate svc rbf and rbf_sampler for CASH_OUT
Previously I wanted to use the RBF sampler (Random Kitchen Sinks) as an
approximation to the SVC + RBF kernel for CASH_OUT. But it performs very
weirdly that for a research valuation, it seems not well comparable.

So I decided to change the training strategy for SVC + RBF on CASH_OUT
by performing stratified downsampling on the CASH_OUT transactions so
that the training size is now similar to the size of the TRANSFER dataset -
at about 300k rows.

Training is now much more feasible with my Mac not dying every single
time.

It will still be interesting to see RBF sampler results on both the
TRANSFER and CASH_OUT dataset though.